chore: move question options to new element data
