+++ packages/graphql/src/schema/practiceQuiz.ts
@@ -5,9 +5,9 @@ import {
 } from '@klicker-uzh/types'
 import builder from '../builder.js'
 import { CourseRef, type ICourse } from './course.js'
+import { ElementType } from './elementData.js'
 import { IStackFeedback } from './evaluation.js'
 import { ElementInstanceRef, InstanceEvaluation } from './question.js'
-import { ElementType } from './questionData.js'
 
 export const ElementOrderType = builder.enumType('ElementOrderType', {
   values: Object.values(DB.ElementOrderType),
